refactor: move to ass.nerdy.autosniper package
