+++ src/main/java/ass/nerdy/autosniper/modules/cF.java
@@ -1,23 +1,19 @@
-package com.bedwars.ntils.modules;
+package ass.nerdy.autosniper.modules;
 
+import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
+
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 
-public class cL {
+public class cF {
     private static final String CONFIG_FILE_PATH = "config/autosniper.json";
-    private JsonObject cJ;
 
-    public cL() {
-        cJ = new JsonObject();
-        initializeConfig();
-    }
-
-    private void initializeConfig() {
+    public static void InitConfig() {
         File configFile = new File(CONFIG_FILE_PATH);
 
         if (!configFile.exists()) {
@@ -38,31 +34,34 @@ public class cL {
             try (FileReader reader = new FileReader(configFile)) {
                 JsonParser parser = new JsonParser();
                 JsonElement element = parser.parse(reader);
-                cJ = element.getAsJsonObject();
+                JsonObject eC = element.getAsJsonObject();
+
+                if (!eC.has("autoRqEnabled")) {
+                    eC.addProperty("autoRqEnabled", false);
+                }
+                if (!eC.has("playerCheckEnabled")) {
+                    eC.addProperty("playerCheckEnabled", false);
+                }
+                if (!eC.has("hudVisible")) {
+                    eC.addProperty("hudVisible", false);
+                }
+                if (!eC.has("fkdrValue")) {
+                    eC.addProperty("fkdrValue", 20);
+                }
+                if (!eC.has("autoRqCommand")) {
+                    eC.addProperty("autoRqCommand", "");
+                }
+                if (!eC.has("apiKey")) {
+                    eC.addProperty("apiKey", "");
+                }
+
+                try (FileWriter writer = new FileWriter(configFile)) {
+                    writer.write(eC.toString());
+                }
+
             } catch (IOException e) {
                 e.printStackTrace();
             }
-        }
-    }
-
-    public boolean gB(String key, boolean dV) {
-        return cJ.has(key) ? cJ.get(key).getAsBoolean() : dV;
-    }
-
-    public String gS(String key, String sdV) {
-        return cJ.has(key) ? cJ.get(key).getAsString() : sdV;
-    }
-
-    public double gD(String key, double ddV) {
-        return cJ.has(key) ? cJ.get(key).getAsDouble() : ddV;
-    }
-
-    public void saveConfig() {
-        File configFile = new File(CONFIG_FILE_PATH);
-        try (FileWriter writer = new FileWriter(configFile)) {
-            writer.write(cJ.toString());
-        } catch (IOException e) {
-            e.printStackTrace();
         }
     }
 }
