+++ src/main/java/ass/nerdy/autosniper/modules/cL.java
@@ -1,19 +1,23 @@
-package com.bedwars.ntils.modules;
+package ass.nerdy.autosniper.modules;
 
-import com.google.gson.Gson;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
-
 import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.IOException;
 
-public class cF {
+public class cL {
     private static final String CONFIG_FILE_PATH = "config/autosniper.json";
+    private JsonObject cJ;
 
-    public static void InitConfig() {
+    public cL() {
+        cJ = new JsonObject();
+        initializeConfig();
+    }
+
+    private void initializeConfig() {
         File configFile = new File(CONFIG_FILE_PATH);
 
         if (!configFile.exists()) {
@@ -34,34 +38,31 @@ public class cF {
             try (FileReader reader = new FileReader(configFile)) {
                 JsonParser parser = new JsonParser();
                 JsonElement element = parser.parse(reader);
-                JsonObject eC = element.getAsJsonObject();
-
-                if (!eC.has("autoRqEnabled")) {
-                    eC.addProperty("autoRqEnabled", false);
-                }
-                if (!eC.has("playerCheckEnabled")) {
-                    eC.addProperty("playerCheckEnabled", false);
-                }
-                if (!eC.has("hudVisible")) {
-                    eC.addProperty("hudVisible", false);
-                }
-                if (!eC.has("fkdrValue")) {
-                    eC.addProperty("fkdrValue", 20);
-                }
-                if (!eC.has("autoRqCommand")) {
-                    eC.addProperty("autoRqCommand", "");
-                }
-                if (!eC.has("apiKey")) {
-                    eC.addProperty("apiKey", "");
-                }
-
-                try (FileWriter writer = new FileWriter(configFile)) {
-                    writer.write(eC.toString());
-                }
-
+                cJ = element.getAsJsonObject();
             } catch (IOException e) {
                 e.printStackTrace();
             }
+        }
+    }
+
+    public boolean gB(String key, boolean dV) {
+        return cJ.has(key) ? cJ.get(key).getAsBoolean() : dV;
+    }
+
+    public String gS(String key, String sdV) {
+        return cJ.has(key) ? cJ.get(key).getAsString() : sdV;
+    }
+
+    public double gD(String key, double ddV) {
+        return cJ.has(key) ? cJ.get(key).getAsDouble() : ddV;
+    }
+
+    public void saveConfig() {
+        File configFile = new File(CONFIG_FILE_PATH);
+        try (FileWriter writer = new FileWriter(configFile)) {
+            writer.write(cJ.toString());
+        } catch (IOException e) {
+            e.printStackTrace();
         }
     }
 }
