+++ src/main/java/ass/nerdy/autosniper/modules/commands/fS.java
@@ -1,6 +1,6 @@
-package com.bedwars.ntils.modules.commands;
+package ass.nerdy.autosniper.modules.commands;
 
-import com.bedwars.ntils.modules.pC;
+import ass.nerdy.autosniper.modules.pC;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import net.minecraft.command.CommandBase;
@@ -12,52 +12,48 @@ import java.io.File;
 import java.io.FileWriter;
 import java.io.FileReader;
 import java.util.HashMap;
-import java.util.Map;
 
-public class aqM extends CommandBase {
+public class fS extends CommandBase {
     private final pC checker;
-    private final Map<String, String> modeMap = new HashMap<>();
     private static final File CONFIG_FILE = new File("config/autosniper.json");
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
-    final String prefix = EnumChatFormatting.GRAY + "[" + EnumChatFormatting.LIGHT_PURPLE + "N" + EnumChatFormatting.GRAY + "] ";
+    private final String prefix = EnumChatFormatting.GRAY + "[" + EnumChatFormatting.LIGHT_PURPLE + "N" + EnumChatFormatting.GRAY + "] ";
 
-    public aqM(pC checker) {
+    public fS(pC checker) {
         this.checker = checker;
-        modeMap.put("1s", "/play bedwars_eight_one");
-        modeMap.put("2s", "/play bedwars_eight_two");
-        modeMap.put("3s", "/play bedwars_four_three");
-        modeMap.put("4s", "/play bedwars_four_four");
-        modeMap.put("4v4", "/play bedwars_two_four");
     }
 
     @Override
     public String getCommandName() {
-        return "mode";
+        return "minfkdr";
     }
 
     @Override
     public String getCommandUsage(ICommandSender sender) {
-        return "/mode [mode]";
+        return "/minfkdr [fkdr value]";
     }
 
     @Override
     public void processCommand(ICommandSender sender, String[] args) {
         if (args.length == 0) {
-            sender.addChatMessage(new ChatComponentText(prefix + EnumChatFormatting.RED + "You must provide a mode to set!"));
+            sender.addChatMessage(new ChatComponentText(prefix + EnumChatFormatting.RED + "You must provide an FKDR value to set!"));
             return;
         }
 
-        String input = args[0].toLowerCase();
-        String command = modeMap.getOrDefault(input, String.join(" ", args));
+        try {
+            double fkdrValue = Double.parseDouble(args[0]);
 
-        checker.setAutoRqCommand(command);
+            JsonFKDR(fkdrValue);
 
-        JsonAutoRq(command);
+            checker.setFkdrValue(String.valueOf(fkdrValue));
 
-        sender.addChatMessage(new ChatComponentText(prefix + EnumChatFormatting.GREEN + "Auto-RQ mode set to: " + EnumChatFormatting.AQUA + command));
+            sender.addChatMessage(new ChatComponentText(prefix + EnumChatFormatting.GREEN + "Minimum FKDR set to: " + EnumChatFormatting.AQUA + fkdrValue));
+        } catch (NumberFormatException e) {
+            sender.addChatMessage(new ChatComponentText(prefix + EnumChatFormatting.RED + "Invalid number. Please enter a valid FKDR."));
+        }
     }
 
-    private void JsonAutoRq(String command) {
+    private void JsonFKDR(double fkdrValue) {
         try {
             if (!CONFIG_FILE.getParentFile().exists()) {
                 CONFIG_FILE.getParentFile().mkdirs();
@@ -70,7 +66,7 @@ public class aqM extends CommandBase {
                 }
             }
 
-            config.put("autoRqCommand", command);
+            config.put("fkdrValue", fkdrValue);
 
             try (FileWriter writer = new FileWriter(CONFIG_FILE)) {
                 GSON.toJson(config, writer);
